# src/swagger.yaml
openapi: 3.0.0
info:
  title: Web Api to Product Stock
  description: Simple api to show how to manager the product stock in many places
  version: 0.1.0
servers:
  - url: http://localhost:3000/v1
    description: base url of the api 

paths:
  /groups:
    get:
      tags:
        - Product Groups
      summary: Product Categories
      responses:
        200:
          content:
            application/json:
                type: []
    post:
      tags: 
        - Product Groups
      summary: Add new Product Categories
      requestBody:
        required: true
        content:
          application/json:
            schema: 
              type: object
              properties:
                name:
                  type: string
      responses:
        202:
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  detail: 
                    type: string
        401:
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: 
                    type: string
  /groups/{productGroupId}:
    delete:
      tags:
        - Product Groups
      parameters:
        - name: productGroupId
          in: path
          required: true
          schema:
           type: string
      responses:
        200:
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: 
                    type: string
                    example: product deleted
        404:
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: 
                    type: string
                    example: product not exist
  /product:
    get:
      tags:
        - Product
      responses:
        200:
          content:
            application/json:
                type: array
    post:
      tags:
        - Product
      requestBody:
        required: true
        content:
          application/json:
            schema: 
              type: object
              properties:
                name:
                  type: string
                productGroupId:
                  type: string
      responses:
        202:
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: 
                    type: string
                    example: product deleted
                  detail:
                    type: string
                    example: created
        400:
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: 
                    type: string
                    example: empty
        403:
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: 
                    type: string
                    example: product exist
                  detail:
                    type: string
        500:
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: 
                    type: string
                    example: internal error
                  detail:
                    type: string
  /product/{productId}:
    delete:
      tags:
        - Product
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: 
                    type: string
                    example: product deleted
        404:
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: 
                    type: string
                    example: product not exist

  /stock:
    get:
      tags:
        - Product Stock
      responses:
        200:
          content:
            application/json:
              type: object
    post:
      tags:
        - Product Stock
      requestBody:
        required: true
        content:
          application/json:
            schema: 
              type: object
              properties:
                productId:
                  type: string
                quantity:
                  type: number
                  example: 1
                lote:
                  type: string
                date:
                  type: object
      responses:
        202:
          content:
            application/json:
              type: object

  /stock/{productId}:
    get:
      tags:
        - Product Stock
      parameters:
        - name: productId 
          in: path
          required: true
          schema:
            type: string
      responses:
        202:
          content:
            application/json:
              type: object
    delete:
      tags:
        - Product Stock
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          application/json:
            type: object
  /stock/out/{stockId}:
    post:
      tags:
        - Product Stock
      parameters:
        - name: stockId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema: 
              type: object
              properties:
                lote:
                  type: string
                quantity:
                  type: number
                  example: 1
                date:
                  type: object
      responses:
        202:
          applicatio/json:
            type: object
  /stock/{stockId}/stock:
    get:
      tags:
        - Product Stock
      parameters:
        - name: stockId
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          application/json:
            type: object
  /moviments:
    get:
      tags:
        - Product Moviments
      responses:
        200:
          content:
            application/json:
              type: object
  /moviments/{productStockId}:
    get:
      tags:
        - Product Moviments
      parameters:
        - name: productStockId
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          content:
            application/json:
              type: object